Inverse rotation directions
The expected rotations are counter-intuitive, but they are required.
This patch therefore inverts "droite" to return a counter-clockwise
rotation and "gauche" to return a clockwise rotation.

An illustration of the change would be rotating left or right by
grabbing the grids at its top and moving it into the specified
direction.

Closes: https://github.com/nc0fr/conception-application/issues/39
Signed-off-by: Nicolas Paul <n@nc0.fr>