Fix Nim's state not updating
The issue was caused by JeuNim having its own etatPartie property,
which was updated internally by the game, instead of updating the one
on the Jeu class.
This caused to create a mismatch between the game's state (for Nim, the
game was over) and the controller (which was seeing Jeu's etatPartie,
still being on EtatPartie.EN_COURS).

Signed-off-by: Nicolas Paul <n@nc0.fr>